enforce comma delimitation of --order arguments
Before, if the first value of --order did not contain the predefined
separator character (the comma), additional arguments were consumed from
the command line as values for --order. As the chain filename is a
positional argument, disable this behavior, ensuring that --order
only takes the one argument following it.

This does not break the behavior of concatenating the values if --order
is passed multiple times.